Move MembershipTier enum to 'loyalty' package
